+++ lib/EasyPost/Claim.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace EasyPost;
+
+/**
+ * @package EasyPost
+ * @property string $id
+ * @property string $approved_amount
+ * @property string[] $attachments
+ * @property string $check_delivery_address
+ * @property string $contact_email
+ * @property string $description
+ * @property object $history
+ * @property string $insurance_amount
+ * @property string $insurance_id
+ * @property string $payment_method
+ * @property string $recipient_name
+ * @property string $requested_amount
+ * @property string $salvage_value
+ * @property string $shipment_id
+ * @property string $status
+ * @property string $status_detail
+ * @property string $status_timestamp
+ * @property string $tracking_code
+ * @property string $type
+ */
+class Claim extends EasyPostObject
+{
+}